requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
cycler==0.11.0
efficientnet-pytorch==0.7.1
Flask==2.1.1
fonttools==4.31.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
threadpoolctl==3.1.0
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
typing-extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
